Extract ElementConnection and ElementProperty
